Added a new temporary task print_debug as a way to centralize all printing into one slow and low-priority task so the others don't get starved. We'll see what happens here; it's just a small patch work for what might be something far more sinister under the hood (perhaps a fake esp32 clone, but that's far-fetched.)
Just comment the #define PRINT_DBG to turn off all console logging when
putting into prod, thanks.

Modified the motors_control function a bit, so that it doesn't confuse
the bot when the error is too high. Hard clamped to +- 90 degrees so
that it's maximum delta should be turning hard left/right. Had to do
this as a consequence of trigo functions, they're not linear.

ultrasonic still not touched yet ;-;