Oh my god fucking finally the ultrasonic sensor finally works
I have a faint idea that interrupts are only allowed for one type of
trigger. In other words, it either has to be a falling or rising edge,
and it can't be both unless i implement a "level change" trigger. Will
consider that another time.

This is just part one of the ultrasonic patch. Will change it out soon.