Added the 'perfect pickup' feature.
Essentially the bot just adjusts itself until it's within a certain
'sweetspot' for collecting trash. Idk if these numbers will work,
testing is really needed for this.

Also timing adjustments for turning and going straight for the obstacle
avoidance routine. This happened in the last commit i think actually,
but it's worth nothing it rn here.

Also made the manual_motor_command more precise when you want a specific
straight line speed. Just made it so because i'm so fucking lazy adding
in the if-else logic checking for if the straight line speed is
negative. idk man